Local Vector Hasher - fixes clang issue
